reworked lots of stuff
the names of things have been redone, mainly so that we can make include
scripts which simply compile multiple scripts together for cleaner
looking defines.  reworked whole project so that it takes on more of a
WYSIWYG feel, as the example project of this calibur can definitely
display the full power of the system.  some debug text displayed below
widget. working on drag and drop functionality next.  added in side
panel Frame Collection to demonstrate the multiple Frame Collection
management abilities.  need to make exmaples of manipulating those as
well.

new data editors planned: dropdown list.  this is probably the next one
to be made.