reworked project more
renamed a bunch of files, trying to clean up the layout of the project.
I am also trying to implement it in a more developer friendly manor.
FrameCollections and Frames should be inheritable, allowing for
customization of their functionality.  FrameCollections will eventually
be used to design more variety of displays, like tabs.